+++ starsky/starsky.foundation.platform/Models/Kestrel/KestrelContainerEndpoints.cs
@@ -1,0 +1,9 @@
+#nullable enable
+namespace starsky.foundation.platform.Models.Kestrel;
+
+public sealed class KestrelContainerEndpoints
+{
+	public KestrelContainerEndpointsUrl? Https { get; set; }
+	public KestrelContainerEndpointsUrl? Http { get; set; }
+
+}